Minor Fixes, Delnar's Skills Remake as Custom Ruleset
Fixed a regression where the karma value calculator would ignore special attributes.

Application Changes:

- Fixed an issue where skill group names would get cut off if they are too long.
- Added the ability for talent priority entries to specify the possibly pool of skills or skillgroups to choose from with an xpath filter.
- Fixed an issue where modifying the load order of a custom data directory after modifying whether it is enabled or not would reset any changes made to whether it was enabled or not.

Data Changes:

- Fixed an issue where Adepts could select Resonance Active skills under priority generation.
- Added a preliminary version of Delnar's Skills Remake's 5th Edition-Like variant as a custom ruleset.
- Fixed an issue where College Education from Run Faster was not giving a 50% discount on karma spent on Academic knowledge skills at character creation.
- Fixed a bug where the Illusionist mastery quality could not be taken with a Spellcasting rank of 4 or 5 if the character had the Illusion specialization.
- Fixed a bug where the Practiced Alchemist mastery quality would not have lower prerequisites for characters with the Islam tradition.
- Fixed a bug where the Altmayr Black Moon would function off of Semi-Automatic specialization instead of Revolvers.

+++ Chummer/frmOptions.cs
@@ -784,6 +784,7 @@ namespace Chummer
 
         private void PopulateCustomDataDirectoryTreeView()
         {
+            object objOldSelected = treCustomDataDirectories.SelectedNode?.Tag;
             if (_lstCustomDataDirectoryInfos.Count != treCustomDataDirectories.Nodes.Count)
             {
                 treCustomDataDirectories.Nodes.Clear();
@@ -810,6 +811,9 @@ namespace Chummer
                     objLoopNode.Checked = objLoopInfo.Enabled;
                 }
             }
+
+            if (objOldSelected != null)
+                treCustomDataDirectories.SelectedNode = treCustomDataDirectories.FindNodeByTag(objOldSelected);
         }
 
         /// <summary>
@@ -1028,25 +1032,41 @@ namespace Chummer
                 }
 
                 // Save the Custom Data Directory Info.
-                if (objRegistry.OpenSubKey("CustomDataDirectory") != null)
-                    objRegistry.DeleteSubKeyTree("CustomDataDirectory");
-                RegistryKey objCustomDataDirectoryRegistry = objRegistry.CreateSubKey("CustomDataDirectory");
-                if (objCustomDataDirectoryRegistry != null)
+                bool blnDoCustomDataDirectoryRefresh = _lstCustomDataDirectoryInfos.Count != GlobalOptions.CustomDataDirectoryInfo.Count;
+                if (!blnDoCustomDataDirectoryRefresh)
                 {
                     for (int i = 0; i < _lstCustomDataDirectoryInfos.Count; ++i)
                     {
-                        CustomDataDirectoryInfo objCustomDataDirectory = _lstCustomDataDirectoryInfos[i];
-                        RegistryKey objLoopKey = objCustomDataDirectoryRegistry.CreateSubKey(objCustomDataDirectory.Name);
-                        if (objLoopKey != null)
+                        if (_lstCustomDataDirectoryInfos[i].CompareTo(GlobalOptions.CustomDataDirectoryInfo[i]) != 0)
                         {
-                            objLoopKey.SetValue("Path", objCustomDataDirectory.Path.Replace(Application.StartupPath, "$CHUMMER"));
-                            objLoopKey.SetValue("Enabled", objCustomDataDirectory.Enabled);
-                            objLoopKey.SetValue("LoadOrder", i);
-                            objLoopKey.Close();
+                            blnDoCustomDataDirectoryRefresh = true;
+                            break;
                         }
                     }
+                }
 
-                    objCustomDataDirectoryRegistry.Close();
+                if (blnDoCustomDataDirectoryRefresh)
+                {
+                    if (objRegistry.OpenSubKey("CustomDataDirectory") != null)
+                        objRegistry.DeleteSubKeyTree("CustomDataDirectory");
+                    RegistryKey objCustomDataDirectoryRegistry = objRegistry.CreateSubKey("CustomDataDirectory");
+                    if (objCustomDataDirectoryRegistry != null)
+                    {
+                        for (int i = 0; i < _lstCustomDataDirectoryInfos.Count; ++i)
+                        {
+                            CustomDataDirectoryInfo objCustomDataDirectory = _lstCustomDataDirectoryInfos[i];
+                            RegistryKey objLoopKey = objCustomDataDirectoryRegistry.CreateSubKey(objCustomDataDirectory.Name);
+                            if (objLoopKey != null)
+                            {
+                                objLoopKey.SetValue("Path", objCustomDataDirectory.Path.Replace(Application.StartupPath, "$CHUMMER"));
+                                objLoopKey.SetValue("Enabled", objCustomDataDirectory.Enabled);
+                                objLoopKey.SetValue("LoadOrder", i);
+                                objLoopKey.Close();
+                            }
+                        }
+
+                        objCustomDataDirectoryRegistry.Close();
+                    }
                 }
 
                 objRegistry.Close();
@@ -1718,7 +1738,7 @@ namespace Chummer
                     txtCharacterRosterPath.Text = dlgSelectFolder.SelectedPath;
             }
         }
-
+        
         private void cmdAddCustomDirectory_Click(object sender, EventArgs e)
         {
             // Prompt the user to select a save file to associate with this Contact.
@@ -1876,6 +1896,20 @@ namespace Chummer
             if (!chkPrintFreeExpenses.Enabled)
                 chkPrintFreeExpenses.Checked = true;
             OptionsChanged(sender, e);
+        }
+
+        private void treCustomDataDirectories_AfterCheck(object sender, TreeViewEventArgs e)
+        {
+            TreeNode objNode = e.Node;
+            if (objNode != null)
+            {
+                CustomDataDirectoryInfo objInfoToRemove = _lstCustomDataDirectoryInfos.FirstOrDefault(x => x.Name == objNode.Tag.ToString());
+                if (objInfoToRemove != null)
+                {
+                    objInfoToRemove.Enabled = objNode.Checked;
+                    OptionsChanged(sender, e);
+                }
+            }
         }
     }
 }
